Add disclaimer and Codecov badge
I have noticed that publicizing this tool may be detrimental for individuals in real competitions seeking to benefit from this tool. Upon new information and further reflection, I find it necessary to include a disclaimer that warns cloners and installers that this program is not yet good enough for real competition use.